Added media channel, restructure forum channel, refactor forum tag -> channel tag
