+++ src/Disqord.Core/Entities/Core/Channels/Guild/IChannelTag.cs
@@ -3,9 +3,9 @@ using Disqord.Models;
 namespace Disqord;
 
 /// <summary>
-///     Represents a tag that can be applied to threads in a forum channel.
+///     Represents a tag that can be applied to threads in a forum or media channel.
 /// </summary>
-public interface IForumTag : IIdentifiableEntity, INamableEntity, IJsonUpdatable<ForumTagJsonModel>
+public interface IChannelTag : IIdentifiableEntity, INamableEntity, IJsonUpdatable<ChannelTagJsonModel>
 {
     /// <summary>
     ///     Gets whether this tag can only be applied to threads
