+++ src/Disqord.Rest.Api/Content/Json/CreateGuildChannelJsonRestRequestContent.cs
@@ -49,7 +49,7 @@ public class CreateGuildChannelJsonRestRequestContent : JsonModelRestRequestCont
     public Optional<GuildChannelFlags> Flags;
 
     [JsonProperty("available_tags")]
-    public Optional<ForumTagJsonModel[]> AvailableTags;
+    public Optional<ChannelTagJsonModel[]> AvailableTags;
 
     [JsonProperty("default_reaction_emoji")]
     public Optional<ForumDefaultReactionJsonModel> DefaultReactionEmoji;
